Generate spp_file based on a directory of input rasters.

Generate spp_file based on a directory of input rasters.

create_spp(filename = "filelist.spp", weight = 1, alpha = 1, bqp = 1,
  bqp_p = 1, cellrem = 0.25, spp_file_dir, recursive = FALSE,
  spp_file_pattern = ".+\\.(tif|img)$", override_path = NULL)

Arguments

filename
character string defining the name of the spp file created.

weight
numeric template value for weights.

alpha
numeric template value for alpha values.

bqp
numeric template value for bqp values.

bqp_p
numeric template value for bqp_p values.

cellrem
numeric template value for cellrem values.

spp_file_dir
character path or a vector of paths to target dir.

recursive
Logical defining whether files in spp_file_dir should be listed recursively.

spp_file_pattern
pattern used to match raster files.

override_path
character path used to override the dirpath in input raster file paths. In case recursive = TRUE, then there can be an arbitrary number of subdirectories and override path is used only up until the spp_file_dir. This way the correct subdirectory structure is retained.

Value

invisible(TRUE), functrion is used for side effects.