Create a new Zonation project on file system.

Based on a set of input arguments, creates a new Zonation project on the file system following a particular folder and file layout.

create_zproject(
  name,
  dir,
  variants,
  dat_template_file = NULL,
  spp_template_file = NULL,
  spp_template_dir = NULL,
  overwrite = FALSE,
  debug = FALSE,
  ...
)

Arguments

name
Character string name for the project. A new directory named by name will be creaated in a location specified by dir.

dir
Character string path pointing to a location to be created.

variants
Character vector of names for new variants. Ignored if using an existing project.

dat_template_file
Character path to a dat file template. If no template is specified, uses the template distributed with zonator. Ignored if using an existing project.

spp_template_file
Character path to a spp file template. If this or spp_template_dir are not specified, uses the template distributed with zonator. Ignored if using an existing project.

spp_template_dir
Character path to directory containing biodiversity feature rasters. If this or spp_template_file are not specified, uses the template distributed with zonator. If both are defined, then spp_template_dir overrides. Ignored if using an existing project.

overwrite
logical should existing project be overwritten (default: FALSE).

debug
logical defining if debugging level for logging should be used (default: FALSE).

...
additional arguments passed to create_spp.

Value

Invisible(NULL) .

Note

This function is used only for the intended side-effect of creating a new Zonation project. To load the project as an instance of Zproject-class, see load_zproject.

See also