Feature names of Zonation variant.

Get and set names for analysis features used a given Zonation variant.

Assign feature names for a Zonation variant.

featurenames(x)

featurenames(x) <- value

# S4 method for ZCurvesDataFrame
featurenames(x)

# S4 method for ZCurvesDataFrame,character
featurenames(x) <- value

# S4 method for Zresults
featurenames(x)

# S4 method for Zvariant
featurenames(x)

# S4 method for Zvariant,character
featurenames(x) <- value

Arguments

x
Z* object.

value
character vector of feature names to be assigned. Can be named or not.

Value

Character vector of spp feature names.

Details

Argument x can be an instance of one the following Z* classes:

Zvariant

Zresults

Note

spp features have by default names that are derived from the feature raster file path.

See also

Zvariant-class Zresults-class groupnames groups

Examples

# NOT RUN {
 setup.dir <- system.file("extdata/tutorial/basic", package="zonator")
 tutorial.project <- create_zproject(setup.dir)
 variant.caz <- get_variant(tutorial.project, "01")

 # Feature names for a Zvariant object
 featurenames(variant.caz)

 # Feature names for a Zresults object
 results.caz <- results(variant.caz)
 featurenames(results.caz)
# }