Get group names for a class Zvariant instance.

Get group names for a class Zvariant instance.

Group names can be assigned to a Zvariant or Zresults object. This is a replacement function for variant group names. If the particular variant doesn't use groups the gives a warning.

groupnames(x)

groupnames(x) <- value

# S4 method for ZGroupCurvesDataFrame
groupnames(x)

# S4 method for Zresults
groupnames(x)

# S4 method for Zvariant
groupnames(x)

# S4 method for Zvariant,character
groupnames(x) <- value

Arguments

x
Zvariant object.

value
named character vector.

Value

A character vector containing the group names. If there are no groups, return NA.

Details

All current group codes must be found in the keys, i.e. there can't be missing values. Argument value can, however, contain keys that are not in the current group codes.

See also

Zvariant-class groupnames groups Zvariant-class groupnames Zresults groups