Get performance levels.

Method returns performance levels at specified levels of cell removal for particular features (either for individual features or groups from a Zresults object).

performance(x, pr.lost, features = "all", groups = FALSE, melted = FALSE)

# S4 method for Zresults
performance(x, pr.lost, features = NULL,
  groups = FALSE, melted = FALSE)

Arguments

x
Zresults object.

pr.lost
numeric vector containing the fractions of landscape lost for which the feature/group performance values are wanted (default: 'all').

features
character vector of features names to be extracted. Must match with feature names in curves data

groups
logical indicating whether group curves data should be used (default: FALSE).

melted
logical indicating whether the data.frame returned should be in melted format (default: FALSE)

Value

Data frame containing the curves file data for selected fractions of landscape lost. First column is pr_lost. If feature names are provided and none are viable, return NA.

Note

In the current implementation, only average performance levels for groups are returned.

See also

Zresults-class read_curves read_grp_curves