Saves the current state of an instance of Zvariant-class on the file system. Zvariant object tracks the location of relevant files, but the root path can be changed. If it is not changed, then the current files can be overwritten.

Saves the current state of an instance of Zvariant-class on the file system. Zvariant object tracks the location of relevant files, but the root path can be changed. If it is not changed, then the current files can be overwritten.

save_zvariant(x, dir = "", overwrite = FALSE, debug_msg = FALSE)

# S4 method for Zvariant
save_zvariant(x, dir = "", overwrite = FALSE,
  debug_msg = FALSE)

Arguments

x
Zvariant object.

dir
Character string path to the root directory where the variant is created.

overwrite
Logical indicating whether files should overwritten if they exist.

debug_msg
Logical setting whether extra debugging information should be printed.

Value

Invisible NULL. This method is used only for it's side effects.

See also

Zvariant-class