Get various Zonation legends

Zonation result rank rasters can be displayed in various color schemes.

zlegend(x)

Arguments

x
String character name for the color scheme.

Value

A list color scheme.

Details

Each color scheme is a list with following item:

values:

Value breaks in the rank priority map

labels:

Labels to be used in the map legend

colors:

Colors used for the value classes

Following color schemes are available:

"spectral"

Note

Color schemes are stored in env .options.

Examples

zlegend("spectral")

#> $values
#> [1] 0.00 0.20 0.50 0.75 0.90 0.95 0.98 1.00
#> 
#> $labels
#> [1] "0.00-0.20" "0.20-0.50" "0.50-0.75" "0.75-0.90" "0.90-0.95" "0.95-0.98"
#> [7] "0.98-1.00"
#> 
#> $colors
#> [1] "#2b83ba" "#80bfab" "#c7e8ad" "#ffffbf" "#fdc980" "#f07c4a" "#d7191c"
#>